公开控制帧与消息页面原子并由插件校验RPC输入
