+++ plugin_boundary_baseline.toml
@@ -35,12 +35,6 @@ R1 = [
     "migrations/yoyo/20260909_01_close_empty_wake_responses.py|plugins.wake.request",
 ]
 R2 = [
-    "plugins/message_push/plugin.py|agent.control.frame_book",
-    "plugins/message_push/restart.py|agent.control.frame_book",
-    "plugins/programmatic/control.py|agent.control.frame_book",
-    "plugins/programmatic/control.py|agent.control.protocol.models",
-    "plugins/programmatic/plugin.py|agent.control.frame_book",
-    "plugins/workbench_ui/dashboard.py|infra.channels.message_view",
 ]
 R3 = [
 ]
