+++ scripts/plugin_boundary.py
@@ -62,6 +62,7 @@ PLUGIN_ALLOWED_MODULES = frozenset({
     "agent.plugin_composition.claims",
     "agent.plugin_composition.commands",
     "agent.plugin_composition.context",
+    "agent.plugin_composition.control_frames",
     "agent.plugin_composition.credentials",
     "agent.plugin_composition.dashboard",
     "agent.plugin_composition.deliveries",
@@ -98,6 +99,7 @@ PLUGIN_ALLOWED_MODULES = frozenset({
     "agent.plugin_contracts",
     "agent.plugin_contracts.json_store",
     "agent.plugin_contracts.message",
+    "agent.plugin_contracts.message_view",
     "core.common.diagnostic_log",
     "core.error_context",
     "core.net.http",
